Move redirectQuery() from lookup-search.js to api/queries.js
