+++ ui/__tests__/lookup-search.test.js
@@ -1,4 +1,4 @@
-import { makeOptionLoader, redirectQuery, search } from '../lookup-search';
+import { makeOptionLoader, search } from '../lookup-search';
 import api from '../util/api/endpoints';
 
 jest.mock('../util/api/endpoints');
@@ -35,17 +35,5 @@ describe('makeOptionLoader()', () => {
         options: [{ value: 4, label: 'four' }, { value: 9, label: 'nine' }],
       });
     });
-  });
-});
-
-describe('redirectQuery()', () => {
-  it('updates an empty query', () => {
-    const result = redirectQuery({ some: 'thing' }, 'myParam', 3);
-    expect(result).toEqual({ some: 'thing', myParam: '3' });
-  });
-  it('updates a populated query', () => {
-    const query = { some: 'thing', myParam: '1,7,9' };
-    const result = redirectQuery(query, 'myParam', 3);
-    expect(result).toEqual({ some: 'thing', myParam: '1,7,9,3' });
   });
 });
