+++ ui/__tests__/util/api/queries.test.js
@@ -1,8 +1,12 @@
-import { cleanSearchParams, formatIssuance, homepageData } from '../queries';
-import api from '../api';
+import {
+  cleanSearchParams,
+  formatIssuance,
+  homepageData,
+  redirectQuery,
+} from '../../../util/api/queries';
+import api from '../../../util/api/endpoints';
 
-jest.mock('../api');
-
+jest.mock('../../../util/api/endpoints');
 
 describe('formatIssuance', () => {
   it('handles reasonable input', () => {
@@ -19,22 +23,21 @@ describe('homepageData', () => {
   it('makes the correct request', () => {
     api.topics.fetchResults.mockImplementationOnce(() => Promise.resolve([]));
     return homepageData().then(() => {
-      expect(api.policies.fetchResults).toHaveBeenCalledWith(
-        { ordering: '-issuance' });
+      expect(api.policies.fetchResults).toHaveBeenCalledWith({ ordering: '-issuance' });
     });
   });
   it('trims to the appropriate length', () => {
     api.topics.fetchResults.mockImplementationOnce(() =>
-      Promise.resolve([1, 2, 3, 4, 5, 6, 7].map(id => ({ id }))));
+      Promise.resolve([1, 2, 3, 4, 5, 6, 7].map(id => ({ id }))),
+    );
     return homepageData().then(({ recentPolicies }) => {
       expect(recentPolicies.map(r => r.id)).toEqual([1, 2, 3, 4]);
     });
   });
   it('formats issuance date', () => {
-    api.topics.fetchResults.mockImplementationOnce(() => Promise.resolve([
-      { issuance: '2002-03-04' },
-      { issuance: '2020-11-10' },
-    ]));
+    api.topics.fetchResults.mockImplementationOnce(() =>
+      Promise.resolve([{ issuance: '2002-03-04' }, { issuance: '2020-11-10' }]),
+    );
     return homepageData().then(({ recentPolicies }) => {
       expect(recentPolicies.map(r => r.issuance_pretty)).toEqual([
         'March 4, 2002',
@@ -89,5 +92,17 @@ describe('cleanSearchParams()', () => {
       lookup: 'topics',
       page: '1',
     });
+  });
+});
+
+describe('redirectQuery()', () => {
+  it('updates an empty query', () => {
+    const result = redirectQuery({ some: 'thing' }, 'myParam', 3);
+    expect(result).toEqual({ some: 'thing', myParam: '3' });
+  });
+  it('updates a populated query', () => {
+    const query = { some: 'thing', myParam: '1,7,9' };
+    const result = redirectQuery(query, 'myParam', 3);
+    expect(result).toEqual({ some: 'thing', myParam: '1,7,9,3' });
   });
 });
