+++ ui/util/api/queries.js
@@ -113,3 +113,19 @@ export async function searchRedirectData({ query }) {
   const pagedEntries = await search(userParams.lookup, userParams.q, userParams.page);
   return { pagedEntries, userParams };
 }
+
+/*
+ * Mix in the idToInsert into the original request parameters.
+ */
+export function redirectQuery(query, insertParam, idToInsert) {
+  const result = Object.assign({}, query);
+  const ids = (result[insertParam] || '').split(',').filter(i => i.length > 0);
+  delete result.page;
+
+  if (!ids.includes(idToInsert)) {
+    ids.push(idToInsert);
+  }
+  result[insertParam] = ids.join(',');
+
+  return result;
+}
